[MOVE]/#8 폴더 response body 형식 함수 모듈화
